Add files via upload
ExoMedia is an audio/video playback library for Android built on top of the ExoPlayer with straightforward APIs and integrations. This library focuses on quick setup, handling common audio and video playback needs while also providing extensibility for more custom use cases.

ExoMedia vs ExoPlayer
The ExoPlayer is an advanced media player for Android that is highly customizable, however that comes at the cost of a more complex setup and configuration process. This customizability is great when it's needed however can be daunting when you need to play a simple audio or video file.

ExoMedia is a more high-level abstraction of media playback that abstracts some of the customizability provided by the ExoPlayer into simple functions and callbacks, keeping the required configuration to a minimum.